requirements.txt
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.141.0
boto3
